+++ CHANGELOG.md
@@ -5,7 +5,8 @@
 
 * `NextflowVdsl3Platform`: Allow passing parameters with multiplicity > 1 from Nextflow CLI.
 
-* `Main`: Add `--cliexport` option which outputs the internal cli construction information to console. This is to be used to automate populating the documentation website.
+* `Main`: Added `viash --cli_export` which outputs the internal cli construction information 
+  to console. This is to be used to automate populating the documentation website.
 
 * `viash ns`: Display overview results with X amount failed & Y amount successfully built.
