+++ src/main/scala/com/dataintuitive/viash/Main.scala
@@ -21,6 +21,7 @@ import java.nio.file.{Files, Path, Paths}
 import java.nio.file.attribute.BasicFileAttributes
 import config.Config
 import helpers.Scala._
+import cli.{CLIConf, CLIExport, ViashCommand, ViashNs}
 
 import scala.collection.JavaConverters
 import java.io.File
